Sorted files, Better Limbs, Fixed VAO/VBO Bug.
Sorted files to have all separate-files be included in the asset folder to be automatically put into the build folders. Made better limb estimation, using cubes instead of lines now. Fixed a bug with vertex arrays that caused multiple attributes to be copied when reusing a shader in the Renderable.